lsps2: framework for scid management
There's a background watcher that scans the lightning node for in-use scids. These scids are cached in memory.
There's a table containing the scids in use for the LSP.

`ScidService.ReserveNewScid` attempts to reserve an unused scid. It tries a few times, just in case it runs into a collision.

To tie this PR together into something working:
- Initialize a `lightning.ScidCache`
- Run `LightningClient.WatchScids` on either the CLN or LND client
- Initialize a `jit.ScidCleaner` and start it
- Initialize a `jit.ScidService`
- Call `ScidService.ReserveNewScid` to reserve a scid