min/max payment size in fee params in lsps2
This change splits the logic for fee params generation between lsps2 and the
grpc protocol, because it adds two fields to the opening fee params. These
fields would otherwise not be able to be added without breaking existing
clients, or having to trust clients to pass the right parameters. Making these
parameters available only in lsps2 gives a clear upgrade path.